+++ src/components/hooks/api/usePrefectures.ts
@@ -3,7 +3,7 @@ import { PrefList } from "@/types/pref"
 import useSWR from "swr"
 import { fetcher } from "@/libs/swr"
 import { PrefecturesResponse200, ResasResponseError } from "@/types/resas"
-import { useResasError } from "@/components/hooks/useResasError"
+import { useResasError } from "@/components/hooks/api/useResasError"
 
 type State = {
   prefList: PrefList
