+++ src/components/hooks/api/useResasError.test.ts
@@ -1,6 +1,6 @@
 import toast from "react-hot-toast"
 import { renderHook } from "@testing-library/react"
-import { useResasError } from "@/components/hooks/useResasError"
+import { useResasError } from "@/components/hooks/api/useResasError"
 import {
   resasResponse400Error,
   resasResponse403Error,
